deal with escaped strings, and produce warnings
This patch has two new features:

* In the Oracle format, it's possible to escape a
character, like so:

    'YYYY-MM-DD"T"HH24:MI:SS'

right now, oracle2posix leaves the quotes in the string,
which causes chagrin for the formatter. The patch
recognize quoted parts of the string, and strip the quotes
in the output of oracle2posix.

* If an Oracle format doesn't have a posix equivalent, oracle2posix
currently silently leave it as is. I've left the behavior as it is,
but added a warning, as this is sure to cause the formatter to
fail down the line.